feat: introduce unified websocket client
